refactor : enum 타입 파일들 위치 변경
