+++ src/main/java/com/capstone/moa/entity/enums/Interest.java
@@ -1,4 +1,4 @@
-package com.capstone.moa.entity;
+package com.capstone.moa.entity.enums;
 
 public enum Interest {
     WEB,
