+++ src/main/java/com/capstone/moa/repository/GroupMemberRepository.java
@@ -2,7 +2,7 @@ package com.capstone.moa.repository;
 
 import com.capstone.moa.entity.Group;
 import com.capstone.moa.entity.GroupMember;
-import com.capstone.moa.entity.GroupRole;
+import com.capstone.moa.entity.enums.GroupRole;
 import com.capstone.moa.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
