+++ src/main/resources/templates/post/view.html
@@ -58,7 +58,7 @@
                 <form th:action="@{/posts/{postId}(postId=${postDetail.postId()})}" method="post">
                     <div class="v104_2684">
                         <div class="comment_profile_image"></div>
-                        <p><textarea  name="email" hidden="hidden" value="mem1@gmail.com" placeholder="작성자 메일을 적어주세요"></textarea></p>
+                        <p><input  name="email" hidden="hidden" value="mem1@gmail.com" placeholder="작성자 추후 변경 필요"/></p>
                         <p><textarea  class="comment_textarea"  name="content" placeholder="댓글을 작성해보세요"></textarea></p>
                         <div class="v30_1649">
                             <button type="submit" class="button"><span class="v30_1651">작성</span>
